chore(): renommage des packages, ajout du fichier de conf tests et de H2
