+++ .gitignore
@@ -34,3 +34,5 @@ out/
 
 ### VS Code ###
 .vscode/
+
+postgres_data
