+++ src/main/java/com/octo/ajava/AjavaApplication.java
@@ -1,0 +1,12 @@
+package com.octo.ajava;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AjavaApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(AjavaApplication.class, args);
+  }
+}
