--- src/main/java/com/octo/ajava/ajava/AjavaApplication.java
@@ -1,13 +1,0 @@
-package com.octo.ajava.ajava;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AjavaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AjavaApplication.class, args);
-	}
-
-}
