--- src/main/java/com/octo/ajava/ajava/domain/FilmVu.java
@@ -1,13 +1,0 @@
-package com.octo.ajava.ajava.domain;
-
-import java.time.LocalDate;
-import java.util.List;
-import java.util.UUID;
-
-public record FilmVu(
-    UUID id,
-    UUID idUtilisateur,
-    String note,
-    String commentaire
-) {
-}
