--- src/main/java/com/octo/ajava/ajava/domain/Utilisateur.java
@@ -1,11 +1,0 @@
-package com.octo.ajava.ajava.domain;
-
-import java.util.UUID;
-
-public record Utilisateur(
-    UUID id,
-    String nom,
-    String prénom,
-    String motDePasse
-) {
-}
