--- src/main/java/com/octo/ajava/ajava/domain/exceptions/NotFoundException.java
@@ -1,11 +1,0 @@
-package com.octo.ajava.ajava.domain.exceptions;
-
-public class NotFoundException extends RuntimeException{
-    private String resource;
-
-    NotFoundException(String resource){
-        this.resource = resource;
-    }
-
-    public String getResource() {return resource;}
-}
