--- src/main/java/com/octo/ajava/ajava/domain/repositories/UtilisateurRepository.java
@@ -1,7 +1,0 @@
-package com.octo.ajava.ajava.domain.repositories;
-
-import com.octo.ajava.ajava.domain.Utilisateur;
-
-public interface UtilisateurRepository {
-    Utilisateur recupererUtilisateurConnecte();
-}
