--- src/main/java/com/octo/ajava/ajava/infra/errorHandling/GlobalControllerAdvice.java
@@ -1,20 +1,0 @@
-package com.octo.ajava.ajava.infra.errorHandling;
-
-import com.octo.ajava.ajava.domain.exceptions.NotFoundException;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Order(Ordered.HIGHEST_PRECEDENCE)
-@RestControllerAdvice(basePackages = {"com.octo.ajava.ajava"})
-public class GlobalControllerAdvice {
-
-    @ExceptionHandler(NotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public MessageErreurApi handleNotFoundException(NotFoundException exception){
-        return new MessageErreurApi("La ressource demandée "+exception.getResource()+" n'a pas été trouvée");
-    }
-}
