--- src/main/java/com/octo/ajava/ajava/infra/errorHandling/MessageErreurApi.java
@@ -1,3 +1,0 @@
-package com.octo.ajava.ajava.infra.errorHandling;
-
-public record MessageErreurApi(String message) {}
