+++ src/main/java/com/octo/ajava/domain/Film.java
@@ -1,0 +1,8 @@
+package com.octo.ajava.domain;
+
+import java.time.LocalDate;
+import java.util.List;
+import java.util.UUID;
+
+public record Film(
+    UUID id, String titre, String synopsis, List<String> genres, LocalDate dateDeSortie) {}
