+++ src/main/java/com/octo/ajava/domain/FilmVu.java
@@ -1,0 +1,5 @@
+package com.octo.ajava.domain;
+
+import java.util.UUID;
+
+public record FilmVu(UUID id, UUID idUtilisateur, String note, String commentaire) {}
