+++ src/main/java/com/octo/ajava/domain/Utilisateur.java
@@ -1,0 +1,5 @@
+package com.octo.ajava.domain;
+
+import java.util.UUID;
+
+public record Utilisateur(UUID id, String nom, String prénom, String motDePasse) {}
