+++ src/main/java/com/octo/ajava/domain/exceptions/NotFoundException.java
@@ -1,0 +1,13 @@
+package com.octo.ajava.domain.exceptions;
+
+public class NotFoundException extends RuntimeException {
+  private String resource;
+
+  NotFoundException(String resource) {
+    this.resource = resource;
+  }
+
+  public String getResource() {
+    return resource;
+  }
+}
