+++ src/main/java/com/octo/ajava/domain/repositories/FilmVuRepository.java
@@ -1,0 +1,12 @@
+package com.octo.ajava.domain.repositories;
+
+import com.octo.ajava.domain.FilmVu;
+import java.util.List;
+
+public interface FilmVuRepository {
+  FilmVu ajouterUnFilmVu() throws Exception;
+
+  List<FilmVu> recupererMesFilmsVus() throws Exception;
+
+  void supprimerUnFilmVu() throws Exception;
+}
