+++ src/main/java/com/octo/ajava/domain/repositories/UtilisateurRepository.java
@@ -1,0 +1,7 @@
+package com.octo.ajava.domain.repositories;
+
+import com.octo.ajava.domain.Utilisateur;
+
+public interface UtilisateurRepository {
+  Utilisateur recupererUtilisateurConnecte();
+}
