+++ src/main/java/com/octo/ajava/domain/usecases/ChercherDesFilmsUseCase.java
@@ -1,0 +1,20 @@
+package com.octo.ajava.domain.usecases;
+
+import com.octo.ajava.domain.Film;
+import com.octo.ajava.domain.repositories.FilmRepository;
+import java.util.List;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ChercherDesFilmsUseCase {
+
+  private final FilmRepository filmRepository;
+
+  ChercherDesFilmsUseCase(FilmRepository filmRepository) {
+    this.filmRepository = filmRepository;
+  }
+
+  public List<Film> executer() throws Exception {
+    return this.filmRepository.chercherDesFilms();
+  }
+}
