+++ src/main/java/com/octo/ajava/domain/usecases/RecupererMesFilmsVusUseCase.java
@@ -1,0 +1,20 @@
+package com.octo.ajava.domain.usecases;
+
+import com.octo.ajava.domain.FilmVu;
+import com.octo.ajava.domain.repositories.FilmVuRepository;
+import java.util.List;
+import org.springframework.stereotype.Component;
+
+@Component
+public class RecupererMesFilmsVusUseCase {
+
+  private final FilmVuRepository filmVuRepository;
+
+  RecupererMesFilmsVusUseCase(FilmVuRepository filmVuRepository) {
+    this.filmVuRepository = filmVuRepository;
+  }
+
+  public List<FilmVu> executer() throws Exception {
+    return this.filmVuRepository.recupererMesFilmsVus();
+  }
+}
