+++ src/main/java/com/octo/ajava/infra/errorHandling/MessageErreurApi.java
@@ -1,0 +1,3 @@
+package com.octo.ajava.infra.errorHandling;
+
+public record MessageErreurApi(String message) {}
