+++ src/test/java/com/octo/ajava/AjavaApplicationTests.java
@@ -1,0 +1,11 @@
+package com.octo.ajava;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AjavaApplicationTests {
+
+  @Test
+  void contextLoads() {}
+}
