--- src/test/java/com/octo/ajava/ajava/AjavaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.octo.ajava.ajava;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AjavaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
