--- src/test/java/com/octo/ajava/ajava/utils/JsonMapperForTesting.java
@@ -1,18 +1,0 @@
-package com.octo.ajava.ajava.utils;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-
-public class JsonMapperForTesting {
-
-    public static String toJson(Object o) throws JsonProcessingException {
-        var objectMapper = new ObjectMapper()
-                .setSerializationInclusion(JsonInclude.Include.NON_NULL)
-                .registerModule(new JavaTimeModule());
-        objectMapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
-        return objectMapper.writeValueAsString(o);
-    }
-}
