+++ src/test/resources/application.properties
@@ -1,0 +1,11 @@
+########## DB
+spring.datasource.username=admin
+spring.datasource.password=passwd
+spring.datasource.type=com.zaxxer.hikari.HikariDataSource
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.hikari.poolName=Hikari
+spring.datasource.hikari.auto-commit=false
+spring.datasource.hikari.data-source-properties.stringtype=unspecified
+spring.sql.init.platform=postgres
+spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.hibernate.ddl-auto=validate
